requirements.txt
beautifulsoup4==4.12.2
Brotli==1.0.9
brotlipy==0.7.0
bs4==0.0.1
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
fonttools==4.40.0
glib==1.0.0
html5lib==1.1
lxml==4.9.2
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==23.1
pangocffi==0.11.0
pdfkit==1.0.0
Pillow==9.5.0
pip-search==0.0.12
pycairo==1.23.0
pydyf==0.6.0
Pygments==2.15.1
PyPDF2==3.0.1
pyphen==0.14.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
python-dotenv==1.0.0
QDarkStyle==3.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.3.1
rich==13.4.2
six==1.16.0
soupsieve==2.4.1
tinycss2==1.2.1
typing_extensions==4.6.3
webencodings==0.5.1
zopfli==0.2.2
